# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Typing App API",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/quizzes": {
      "get": {
        "summary": "Get Quizzes",
        "operationId": "get_quizzes_quizzes_get",
        "parameters": [
          {
            "name": "eiken_levels",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Eiken Levels"
            }
          },
          {
            "name": "question_types",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Question Types"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tags"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/questions": {
      "get": {
        "summary": "Get Questions",
        "operationId": "get_questions_questions_get",
        "parameters": [
          {
            "name": "eiken_levels",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Eiken Levels"
            }
          },
          {
            "name": "question_types",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Question Types"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tags"
            }
          },
          {
            "name": "include_inactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Inactive"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Post Question",
        "operationId": "post_question_questions_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/questions/{question_id}": {
      "patch": {
        "summary": "Patch Question",
        "operationId": "patch_question_questions__question_id__patch",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Question",
        "operationId": "delete_question_questions__question_id__delete",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Question Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/study-results": {
      "post": {
        "summary": "Post Study Result",
        "operationId": "post_study_result_study_results_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyResultRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyResultRequest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/study-results/latest": {
      "get": {
        "summary": "Get Latest Result",
        "operationId": "get_latest_result_study_results_latest_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyResultRequest"
                }
              }
            }
          }
        }
      }
    },
    "/study-results/summary/today": {
      "get": {
        "summary": "Get Today Summary",
        "operationId": "get_today_summary_study_results_summary_today_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyStudySummaryResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DailyStudySummaryResponse": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "sessions": {
            "type": "integer",
            "title": "Sessions"
          },
          "solvedProblems": {
            "type": "integer",
            "title": "Solvedproblems"
          }
        },
        "type": "object",
        "required": [
          "date",
          "sessions",
          "solvedProblems"
        ],
        "title": "DailyStudySummaryResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "QuestionCreate": {
        "properties": {
          "eiken_level_code": {
            "type": "string",
            "minLength": 1,
            "title": "Eiken Level Code"
          },
          "question_type": {
            "type": "string",
            "enum": [
              "word",
              "sentence"
            ],
            "title": "Question Type"
          },
          "english": {
            "type": "string",
            "minLength": 1,
            "title": "English"
          },
          "japanese": {
            "type": "string",
            "minLength": 1,
            "title": "Japanese"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "eiken_level_code",
          "question_type",
          "english",
          "japanese"
        ],
        "title": "QuestionCreate"
      },
      "QuestionListResponse": {
        "properties": {
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionResponse"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "questions"
        ],
        "title": "QuestionListResponse"
      },
      "QuestionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "word",
              "sentence"
            ],
            "title": "Type"
          },
          "eikenLevel": {
            "type": "string",
            "title": "Eikenlevel"
          },
          "english": {
            "type": "string",
            "title": "English"
          },
          "japanese": {
            "type": "string",
            "title": "Japanese"
          },
          "isActive": {
            "type": "boolean",
            "title": "Isactive"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "eikenLevel",
          "english",
          "japanese",
          "isActive",
          "tags"
        ],
        "title": "QuestionResponse"
      },
      "QuestionUpdate": {
        "properties": {
          "eiken_level_code": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Eiken Level Code"
          },
          "question_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "word",
                  "sentence"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Type"
          },
          "english": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "English"
          },
          "japanese": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Japanese"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "type": "object",
        "title": "QuestionUpdate"
      },
      "QuizListResponse": {
        "properties": {
          "quizzes": {
            "items": {
              "$ref": "#/components/schemas/QuizResponse"
            },
            "type": "array",
            "title": "Quizzes"
          }
        },
        "type": "object",
        "required": [
          "quizzes"
        ],
        "title": "QuizListResponse"
      },
      "QuizResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "word",
              "sentence"
            ],
            "title": "Type"
          },
          "eikenLevel": {
            "type": "string",
            "title": "Eikenlevel"
          },
          "english": {
            "type": "string",
            "title": "English"
          },
          "japanese": {
            "type": "string",
            "title": "Japanese"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "eikenLevel",
          "english",
          "japanese",
          "tags"
        ],
        "title": "QuizResponse"
      },
      "StudyResultRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "learn",
              "review"
            ],
            "title": "Mode"
          },
          "total_questions": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Total Questions"
          },
          "correct_rate": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Correct Rate"
          },
          "mistakes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Mistakes"
          },
          "average_time": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Average Time"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "total_questions",
          "correct_rate",
          "mistakes",
          "average_time",
          "created_at"
        ],
        "title": "StudyResultRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}